+++ shimmer/display/widgets/dialog_window.py
@@ -7,7 +7,7 @@ For example, an "Are you sure" dialog with yes/no answers.
 
 from dataclasses import replace
 from functools import update_wrapper
-from typing import Sequence, Optional
+from typing import Optional, Set
 
 from .question_definition import (
     MultipleChoiceQuestionDefinition,
@@ -20,7 +20,12 @@ from .multiple_choice_buttons import (
 )
 from .button import ButtonDefinition
 from .window import WindowDefinition, Window
-from ..components.mouse_box import MouseEventCallable
+from ..components.mouse_box import MouseClickEventCallable
+
+
+AreYouSure = MultipleChoiceQuestionDefinition(
+    text="Are you sure?", choices=["Yes", "No"], allow_multiple=False,
+)
 
 
 def callback_with_answer_and_close_window(
@@ -29,7 +34,7 @@ def callback_with_answer_and_close_window(
     """Create a callback that closes the window as well as calling the given question callback."""
 
     def inner(
-        currently_selected: Sequence[str], changed_choice: str, choice_state: bool
+        currently_selected: Set[str], changed_choice: str, choice_state: bool
     ) -> None:
         callback(currently_selected, changed_choice, choice_state)
         window.kill()
@@ -41,27 +46,31 @@ def callback_with_answer_and_close_window(
 def create_multiple_choice_question_dialog(
     definition: MultipleChoiceQuestionDefinition,
     on_answer: MultipleChoiceResponseCallback,
-    on_cancel: Optional[QuestionCancelledCallback],
 ) -> Window:
     """
     Create a window containing a multiple choice question.
 
+    If the question allows multiple answers, a submit button will be shown that closes the window.
+    The submit button doe not call any additional callback as each answer choice will have called
+    back already.
+
+    If only a single answer is allowed, the window will close as soon as an option is chosen.
+
     :param definition: Definition of the question.
     :param on_answer: Called when an answer is selected.
-    :param on_cancel: Called when the window is closed.
     """
-    on_close: Optional[MouseEventCallable]
+    on_close: Optional[MouseClickEventCallable]
 
-    def cancel_on_close() -> MouseEventCallable:
+    def cancel_on_close() -> MouseClickEventCallable:
         """Wrapper around the mouse event callback on close that ignores all the arguments."""
 
         def inner(*_, **__):
-            if on_cancel is not None:
-                on_cancel()
+            if definition.on_cancel is not None:
+                definition.on_cancel()
 
         return inner
 
-    if on_cancel is not None:
+    if definition.on_cancel is not None:
         on_close = cancel_on_close()
     else:
         on_close = None
@@ -107,10 +116,5 @@ def create_are_you_sure_dialog(
     :param on_cancel: Method to call if the X button is used instead of answering the question.
     :return: Window containing Yes/No buttons.
     """
-    question_defn = MultipleChoiceQuestionDefinition(
-        text="Are you sure?",
-        choices=["Yes", "No"],
-        allow_multiple=False,
-        on_cancel=on_cancel,
-    )
-    return create_multiple_choice_question_dialog(question_defn, on_answer, on_cancel)
+    question_defn = replace(AreYouSure, on_cancel=on_cancel)
+    return create_multiple_choice_question_dialog(question_defn, on_answer)